+++ includes/EditorLevelController.hpp
@@ -91,9 +91,9 @@ public:
      *
      * @param context commons items for controller
      *
-     * @return uint8_t    id of the new screen controller
+     * @return unsigned short id of the new screen controller
      */
-    uint8_t render(utils::Context& context);
+    unsigned short render(utils::Context& context);
 
 private:
