+++ includes/SerieMainMenuController.hpp
@@ -51,11 +51,9 @@ public:
      *
      * @param context commons items for controller
      *
-     * @return uint8_t   id of the next screen controller
+     * @return unsigned short id of the next screen controller
      */
-    uint8_t render(
-        utils::Context& context
-    );
+    unsigned short render(utils::Context& context);
 
 private:
